+++ tests/unit_tests/definition/test_inference_problem.py
@@ -10,7 +10,7 @@ import numpy as np
 from probeye.definition.inference_problem import InferenceProblem
 from probeye.definition.forward_model import ForwardModelBase
 from probeye.definition.sensor import Sensor
-from probeye.definition.noise_model import NoiseModelBase
+from probeye.definition.likelihood_model import NoiseModelBase
 
 
 class TestProblem(unittest.TestCase):
@@ -21,7 +21,7 @@ class TestProblem(unittest.TestCase):
         self.assertEqual(p.n_latent_prms, 1)
         self.assertEqual(p.latent_prms_dims, [1])
         # add second latent parameter; n_latent_prms should be 2
-        p.add_parameter("b", "noise", prior=("normal", {"loc": 0.0, "scale": 1.0}))
+        p.add_parameter("b", "likelihood", prior=("normal", {"loc": 0.0, "scale": 1.0}))
         self.assertEqual(p.n_latent_prms, 2)
         self.assertEqual(p.latent_prms_dims, [1, 1])
         # check the different properties
@@ -37,8 +37,8 @@ class TestProblem(unittest.TestCase):
         self.assertEqual(set(p.model_prms), {"a"})
         self.assertEqual(p.n_prior_prms, 4)
         self.assertEqual(set(p.prior_prms), {"loc_a", "scale_a", "loc_b", "scale_b"})
-        self.assertEqual(p.n_noise_prms, 1)
-        self.assertEqual(set(p.noise_prms), {"b"})
+        self.assertEqual(p.n_likelihood_prms, 1)
+        self.assertEqual(set(p.likelihood_prms), {"b"})
 
     def test_info(self):
         # simply check that no errors occur when info-method is called; the output that
@@ -46,30 +46,28 @@ class TestProblem(unittest.TestCase):
         p = InferenceProblem("TestProblem")
         p.add_parameter("a", "model", const=1.0)
         p.add_parameter("b", "model", prior=("normal", {"loc": 0, "scale": 1}))
-        p.add_parameter("s", "noise", prior=("normal", {"loc": 0, "scale": 1}))
+        p.add_parameter("s", "likelihood", prior=("normal", {"loc": 0, "scale": 1}))
         sys.stdout = io.StringIO()
         # try out different options
-        p.info(include_experiments=False, tablefmt="presto", check_consistency=False)
-        _ = p.info(
-            include_experiments=False, tablefmt="presto", check_consistency=False
-        )
-        p.info(include_experiments=True, tablefmt="presto", check_consistency=False)
-        p.info(include_experiments=True, tablefmt="plain", check_consistency=False)
+        p.info(tablefmt="presto", check_consistency=False)
+        _ = p.info(tablefmt="presto", check_consistency=False)
+        p.info(tablefmt="presto", check_consistency=False)
+        p.info(tablefmt="plain", check_consistency=False)
         sys.stdout = sys.__stdout__  # reset printout to console
         with self.assertRaises(AssertionError):
             # the problem is not consistent yet (e.g. no forward model defined yet), so
             # the consistency_check will raise an error
-            p.info(include_experiments=True, tablefmt="presto", check_consistency=True)
+            p.info(tablefmt="presto", check_consistency=True)
         # now add the remaining stuff to make to problem consistent
         test_model = ForwardModelBase("b", Sensor("x"), Sensor("y"))
         p.add_forward_model("TestModel", test_model)
-        p.add_noise_model(NoiseModelBase("normal", "s", sensors=Sensor("y")))
         p.add_experiment(
             "Experiment_1", sensor_values={"x": 1, "y": 1}, fwd_model_name="TestModel"
         )
+        p.add_likelihood_model(NoiseModelBase("normal", "s", sensors=Sensor("y")))
         sys.stdout = io.StringIO()
         # now, the consistency_check should not raise an error
-        p.info(include_experiments=True, tablefmt="presto", check_consistency=True)
+        p.info(tablefmt="presto", check_consistency=True)
         sys.stdout = sys.__stdout__  # reset printout to console
 
     def test_str(self):
@@ -77,13 +75,13 @@ class TestProblem(unittest.TestCase):
         p = InferenceProblem("TestProblem")
         p.add_parameter("a", "model", const=1.0)
         p.add_parameter("b", "model", prior=("normal", {"loc": 0, "scale": 1}))
-        p.add_parameter("s", "noise", prior=("normal", {"loc": 0, "scale": 1}))
+        p.add_parameter("s", "likelihood", prior=("normal", {"loc": 0, "scale": 1}))
         test_model = ForwardModelBase("b", Sensor("x"), Sensor("y"))
         p.add_forward_model("TestModel", test_model)
-        p.add_noise_model(NoiseModelBase("normal", "s", sensors=Sensor("y")))
         p.add_experiment(
             "Experiment_1", sensor_values={"x": 1, "y": 1}, fwd_model_name="TestModel"
         )
+        p.add_likelihood_model(NoiseModelBase("normal", "s", sensors=Sensor("y")))
         sys.stdout = io.StringIO()  # redirect output to console
         print(p)
         sys.stdout = sys.__stdout__  # reset printout to console
@@ -93,7 +91,9 @@ class TestProblem(unittest.TestCase):
         # check valid use cases for constant parameters
         p.add_parameter("c", "model", const=1.0, info="info", tex=r"$c$")
         p.add_parameter("loc_a", "prior", const=1.0, info="info", tex=r"$loc_a$")
-        p.add_parameter("sigma_1", "noise", const=1.0, info="info", tex=r"$\sigma_1$")
+        p.add_parameter(
+            "sigma_1", "likelihood", const=1.0, info="info", tex=r"$\sigma_1$"
+        )
         # check valid use cases for latent parameters
         p.add_parameter(
             "b",
@@ -111,7 +111,7 @@ class TestProblem(unittest.TestCase):
         )
         p.add_parameter(
             "sigma_2",
-            "noise",
+            "likelihood",
             info="info",
             tex=r"$\sigma_2$",
             prior=("uniform", {"low": 0.1, "high": 2.0}),
@@ -135,7 +135,7 @@ class TestProblem(unittest.TestCase):
             )
         with self.assertRaises(RuntimeError):
             # adding a parameter with a name that was used before
-            p.add_parameter("c", "noise", const=2.0, info="info", tex=r"$c$")
+            p.add_parameter("c", "likelihood", const=2.0, info="info", tex=r"$c$")
         with self.assertRaises(TypeError):
             # adding a parameter with an invalid prior
             p.add_parameter(
@@ -196,7 +196,7 @@ class TestProblem(unittest.TestCase):
         p = InferenceProblem("TestProblem")
         p.add_parameter("c", "model", const=1.0, info="info", tex=r"$c$")
         p.add_parameter("loc_a", "prior", const=1.0, info="info", tex=r"$loc_a$")
-        p.add_parameter("s", "noise", const=1.0, info="info", tex=r"$\sigma$")
+        p.add_parameter("s", "likelihood", const=1.0, info="info", tex=r"$\sigma$")
         p.add_parameter(
             "scale_a",
             "prior",
@@ -206,7 +206,7 @@ class TestProblem(unittest.TestCase):
         )
         p.add_parameter(
             "sigma",
-            "noise",
+            "likelihood",
             info="info",
             tex=r"$\sigma$",
             prior=("uniform", {"low": 0.1, "high": 2.0}),
@@ -298,6 +298,17 @@ class TestProblem(unittest.TestCase):
             # change to role the parameter already has
             p.change_parameter_role("a", prior=("normal", {"loc": 0, "scale": 1}))
 
+    def test_change_parameter_type(self):
+        p = InferenceProblem("TestProblem")
+        # check change of type from 'model' to 'likelihood'
+        p.add_parameter("a", "model", prior=("normal", {"loc": 0, "scale": 1}))
+        self.assertEqual(p.parameters["a"].type, "model")
+        p.change_parameter_type("a", "likelihood")
+        self.assertEqual(p.parameters["a"].type, "likelihood")
+        # check change of type from 'likelihood' to invalid type 'INVALID'
+        with self.assertRaises(ValueError):
+            p.change_parameter_type("a", "INVALID")
+
     def test_change_parameter_info(self):
         p = InferenceProblem("TestProblem")
         # simple check that the change works and has the expected effect
@@ -367,17 +378,14 @@ class TestProblem(unittest.TestCase):
         with self.assertRaises(AssertionError):
             # no noise models defined yet
             p.check_problem_consistency()
-        # add a noise model
-        p.add_parameter("s", "noise", const=1.0)
-        noise_model = NoiseModelBase("normal", "s", sensors=[Sensor("y")])
-        p.add_noise_model(noise_model)
-        with self.assertRaises(AssertionError):
-            # no experiment_names defined yet
-            p.check_problem_consistency()
-        # add at an experiment
+        # add an experiment
         p.add_experiment(
             "Experiment_1", sensor_values={"x": 1, "y": 1}, fwd_model_name="TestModel"
         )
+        # add a noise model
+        p.add_parameter("s", "likelihood", const=1.0)
+        noise_model = NoiseModelBase("normal", "s", sensors=[Sensor("y")])
+        p.add_likelihood_model(noise_model)
         # now the problem should be consistent
         p.check_problem_consistency()
 
@@ -430,13 +438,6 @@ class TestProblem(unittest.TestCase):
             # forward model's output sensor not provided by experiment
             p.add_experiment(
                 "Experiment_3", sensor_values={"x": 1}, fwd_model_name="TestModel"
-            )
-        with self.assertRaises(RuntimeError):
-            # sensor values with different lengths
-            p.add_experiment(
-                "Experiment_3",
-                fwd_model_name="TestModel",
-                sensor_values={"x": [1, 2], "y": 1},
             )
         # check that sensor_value lists are transformed to numpy arrays
         p.add_experiment(
@@ -595,7 +596,7 @@ class TestProblem(unittest.TestCase):
         p = InferenceProblem("TestProblem")
         p.add_parameter("a", "model", const=1.0)
         p.add_parameter("b", "model", prior=("normal", {"loc": 0, "scale": 1}))
-        p.add_parameter("s", "noise", prior=("normal", {"loc": 0, "scale": 1}))
+        p.add_parameter("s", "likelihood", prior=("normal", {"loc": 0, "scale": 1}))
         # check the model_consistency flag
         with self.assertRaises(AssertionError):
             # the model is not consistent
@@ -624,11 +625,14 @@ class TestProblem(unittest.TestCase):
         # check correct use
         p = InferenceProblem("TestProblem")
         p.add_parameter("a", "model", prior=("normal", {"loc": 0, "scale": 1}))
-        p.add_parameter("s1", "noise", prior=("normal", {"loc": 0, "scale": 1}))
-        p.add_parameter("s2", "noise", prior=("normal", {"loc": 0, "scale": 1}))
-        p.add_parameter("s3", "noise", prior=("normal", {"loc": 0, "scale": 1}))
+        p.add_parameter("s1", "likelihood", prior=("normal", {"loc": 0, "scale": 1}))
+        p.add_parameter("s2", "likelihood", prior=("normal", {"loc": 0, "scale": 1}))
+        p.add_parameter("s3", "likelihood", prior=("normal", {"loc": 0, "scale": 1}))
         test_model = ForwardModelBase("a", Sensor("x"), [Sensor("y1"), Sensor("y2")])
         p.add_forward_model("TestModel", test_model)
+        p.add_experiment(
+            "Exp", fwd_model_name="TestModel", sensor_values={"x": 0, "y1": 0, "y2": 0}
+        )
         noise_model1 = NoiseModelBase(
             "normal", "s1", sensors=[Sensor("y1"), Sensor("y2")]
         )
@@ -636,15 +640,15 @@ class TestProblem(unittest.TestCase):
         noise_model3 = NoiseModelBase(
             "normal", ["s1", "s2", "s3"], sensors=[Sensor("y1")], name="NM3"
         )
-        p.add_noise_model(noise_model1)
-        p.add_noise_model(noise_model2)
-        p.add_noise_model(noise_model3)
+        p.add_likelihood_model(noise_model1)
+        p.add_likelihood_model(noise_model2)
+        p.add_likelihood_model(noise_model3)
         # adding a noise model with similar sensor interface should not error
-        p.add_noise_model(noise_model3)
+        p.add_likelihood_model(noise_model3)
         # check invalid input arguments
         with self.assertRaises(RuntimeError):
             # the given noise model parameter has not been defined
-            p.add_noise_model(
+            p.add_likelihood_model(
                 NoiseModelBase(
                     "normal",
                     "not_existing_parameter",
@@ -657,9 +661,9 @@ class TestProblem(unittest.TestCase):
         p = InferenceProblem("TestProblem")
         # add some parameters
         p.add_parameter("a", "model", prior=("normal", {"loc": 0, "scale": 1}))
-        p.add_parameter("s1", "noise", prior=("normal", {"loc": 0, "scale": 1}))
-        p.add_parameter("s2", "noise", prior=("normal", {"loc": 0, "scale": 1}))
-        p.add_parameter("s3", "noise", prior=("normal", {"loc": 0, "scale": 1}))
+        p.add_parameter("s1", "likelihood", prior=("normal", {"loc": 0, "scale": 1}))
+        p.add_parameter("s2", "likelihood", prior=("normal", {"loc": 0, "scale": 1}))
+        p.add_parameter("s3", "likelihood", prior=("normal", {"loc": 0, "scale": 1}))
         # add some dummy forward models (only the output sensors are important)
         test_model_y1 = ForwardModelBase("a", Sensor("x"), Sensor("y1"))
         test_model_y2 = ForwardModelBase("a", Sensor("x"), Sensor("y2"))
@@ -669,16 +673,6 @@ class TestProblem(unittest.TestCase):
         p.add_forward_model("TestModel_y1", test_model_y1)
         p.add_forward_model("TestModel_y2", test_model_y2)
         p.add_forward_model("TestModel_z1z2", test_model_z1z2)
-        # add some noise models
-        noise_model_y1 = NoiseModelBase(
-            "normal", ["s1", "s2", "s3"], sensors=[Sensor("y1")]
-        )
-        noise_model_y2 = NoiseModelBase("normal", ["s2"], sensors=[Sensor("y2")])
-        noise_model_y1y2 = NoiseModelBase(
-            "normal", "s1", sensors=[Sensor("z1"), Sensor("z2")]
-        )
-        p.add_noise_model(noise_model_y1)
-        p.add_noise_model(noise_model_y2)
         # add some experiments
         p.add_experiment(
             "Exp_y1", sensor_values={"x": 1, "y1": 2}, fwd_model_name="TestModel_y1"
@@ -694,18 +688,25 @@ class TestProblem(unittest.TestCase):
             sensor_values={"x": 7, "z1": 8, "z2": 9},
             fwd_model_name="TestModel_z1z2",
         )
-        # so far, not all noise models have been added, so there are unassigned
-        # experiments; this should lead to an error
-        with self.assertRaises(RuntimeError):
-            p.assign_experiments_to_noise_models()
-        # now we add the missing noise model
-        p.add_noise_model(noise_model_y1y2)
+        # add some noise models
+        noise_model_y1 = NoiseModelBase(
+            "normal", ["s1", "s2", "s3"], sensors=[Sensor("y1")]
+        )
+        noise_model_y2 = NoiseModelBase("normal", ["s2"], sensors=[Sensor("y2")])
+        noise_model_y1y2 = NoiseModelBase(
+            "normal", "s1", sensors=[Sensor("z1"), Sensor("z2")]
+        )
+        p.add_likelihood_model(noise_model_y1, name="l1")
+        p.add_likelihood_model(noise_model_y2, name="l2")
+        p.add_likelihood_model(noise_model_y1y2, name="l3")
         # this is the call that should be tested here
-        p.assign_experiments_to_noise_models()
+        p.check_problem_consistency()
         # now check if all experiments have been assigned correctly
-        self.assertEqual(p.noise_models[0].experiment_names, ["Exp_y1"])
-        self.assertEqual(p.noise_models[1].experiment_names, ["Exp_y2_1", "Exp_y2_2"])
-        self.assertEqual(p.noise_models[2].experiment_names, ["Exp_z1z2"])
+        self.assertEqual(p.likelihood_models["l1"].experiment_names, ["Exp_y1"])
+        self.assertEqual(
+            p.likelihood_models["l2"].experiment_names, ["Exp_y2_1", "Exp_y2_2"]
+        )
+        self.assertEqual(p.likelihood_models["l3"].experiment_names, ["Exp_z1z2"])
 
     def test_transform_experimental_data(self):
         # check correct use
@@ -719,7 +720,9 @@ class TestProblem(unittest.TestCase):
             sensor_values={"x": [1, 2], "y": [1, 3]},
         )
         # apply a simple power function to the experimental data
-        p_copy = p.transform_experimental_data(f=np.power, args=([2, 3],), where=True)
+        p_copy = p.transform_experimental_data(
+            func=np.power, args=([2, 3],), where=True
+        )
         x_computed = p_copy.experiments["Experiment_1"]["sensor_values"]["x"]
         x_expected = np.array([1, 8])
         self.assertTrue(